Address comments from Brian

+++ src/Tribe/iCal.php
@@ -606,6 +606,7 @@ class Tribe__Events__iCal {
 	 *
 	 * @since 4.9.4
 	 * @since 6.10.2 Make sure that each time zone definition has its own group.
+	 * @since TBD Adjust the time zone definition to include the DST transitions for a year before and after.
 	 *
 	 * @param array $events An array with all the events.
 	 *
@@ -642,14 +643,14 @@ class Tribe__Events__iCal {
 			 *
 			 * @since TBD
 			 *
-			 * @param int    $years    The number of years to extend before and after event years. Default 3.
+			 * @param int    $years    The number of years to extend before and after event years. Default 1.
 			 * @param string $timezone The timezone identifier (e.g., 'Europe/Berlin').
 			 * @param array  $events   The events being processed for this timezone.
 			 */
-			$extend_years = apply_filters( 'tec_events_ical_timezone_extend_years', 1, $timezone->getName(), $row['events'] );
+			$extend_years = (int) apply_filters( 'tec_events_ical_timezone_extend_years', 1, $timezone->getName(), $row['events'] );
 
 			// Ensure we have a valid positive integer.
-			$extend_years = max( 1, (int) $extend_years );
+			$extend_years = max( 1, $extend_years );
 
 			// Extend the range by the specified number of years in each direction.
 			$extended_start = strtotime( 'first day of january ' . ( $start_year - $extend_years ) );
